Region map rework, Guyana progress, Hoenn's Slateport Harbor
Implemented changes to how the region maps load, to help with bugs. Relies on having the 8th Kanto gym badge to switch through all possible region maps. Also works when flying to a known location in the region.

+++ include/pokenav.h
@@ -17,7 +17,7 @@ struct PokenavMonListItem
 struct PokenavMatchCallEntry
 {
     bool8 isSpecialTrainer;
-    u8 mapSec;
+    u16 mapSec;
     u16 headerId;
 };
